Revert "refactor(core): Use `ActionResolver` in `Dispatcher`. (#56369)"
This reverts commit 4ebd2853fa7a9b47c71ff19b5a6a819502905508.
